為 coutomize 抽象 HotKeyManager 類別
